Moved helper func out of mutation and into helpers for context
